:recycle: refactor: `Bettings` 베팅 금액 업데이트 기능 추가
