+++ README.md
@@ -83,6 +83,8 @@
   - [x] 플레이어 베팅 금액 등록
   - [x] 플레이어 베팅 금액 검색
   - [x] 딜러 베팅 금액 계산
+  - [x] 플레이어 베팅 금액 업데이트
+  - [x] 플레이어 게임 상태 판별
 
 ### ParticipantCardsDto
 - [x] 참여자의 이름, 참여자가 보유한 `Cards`, `Cards`의 총 합
